refactor validation selector setup
