+++ support-frontend/assets/components/stripeCardForm/selectors.ts
@@ -1,36 +1,23 @@
 import type { StripeFormErrors } from 'helpers/redux/checkout/payment/stripe/state';
 import type { ContributionsState } from 'helpers/redux/contributionsStore';
+import { getStripeFormErrors } from 'helpers/redux/selectors/formValidation/paymentValidation';
 
 export type StripeCardFormDisplayErrors = StripeFormErrors & {
 	recaptcha?: string[];
 	zipCode?: string[];
 };
 
-export function getStripeCardFormErrors(state: ContributionsState): {
+export function getDisplayErrors(state: ContributionsState): {
 	errors: StripeCardFormDisplayErrors;
 	showErrors: boolean;
 } {
-	const { errors, showErrors } = state.page.checkoutForm.payment.stripe;
+	const { showErrors } = state.page.checkoutForm.payment.stripe;
 	const recaptchaErrors = state.page.checkoutForm.recaptcha.errors;
-
-	const shouldShowZipCode =
-		state.common.internationalisation.countryId === 'US';
-
-	if (shouldShowZipCode) {
-		const zipCode =
-			state.page.checkoutForm.billingAddress.fields.errorObject?.postCode;
-		return {
-			errors: {
-				...errors,
-				zipCode,
-			},
-			showErrors,
-		};
-	}
+	const formErrors = getStripeFormErrors(state);
 
 	return {
 		errors: {
-			...errors,
+			...formErrors,
 			recaptcha: recaptchaErrors,
 		},
 		showErrors,
