+++ support-frontend/assets/helpers/redux/selectors/formValidation/otherAmountValidation.ts
