+++ support-frontend/assets/helpers/redux/selectors/formValidation/utils.ts
@@ -1,0 +1,7 @@
+export type ErrorCollection = Partial<Record<string, string[]>>;
+
+export function errorCollectionHasErrors(
+	errorCollection: ErrorCollection,
+): boolean {
+	return Object.values(errorCollection).some((errorList) => errorList?.length);
+}
